updated working bottle implementation
Most of it was fixed on spline not much was modified in this code, I think this is just going back to the starting code which is now being used to work all transitions.
Minro bugs need fixing: Hower not working for some reason.
Changes needed:
Need to add buttons like in the can page.